pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-agent"
version = "0.1.0"
description = "AI-powered PDF analysis and reporting system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf==1.24.0",
    "pdfplumber==0.11.0",
    "PyPDF2==3.0.1",
    "pillow==10.2.0",
    "numpy==1.26.3",
    "pandas==2.2.0",
    "pydantic==2.6.0",
    "typing-extensions==4.9.0",
    "python-dotenv==1.0.0",
    "anthropic>=0.39.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
pdf2tasks = "src.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "black==24.1.1",
    "mypy==1.8.0",
    "ruff==0.1.14",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"
